Separating the test case
