+++ core/src/test/scala/org/apache/spark/deploy/master/MasterSuite.scala
@@ -19,33 +19,25 @@ package org.apache.spark.deploy.master
 
 import java.net.{HttpURLConnection, URL}
 import java.util.Date
-import java.util.concurrent.{ConcurrentLinkedQueue, CountDownLatch, TimeUnit}
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.{ConcurrentLinkedQueue, TimeUnit}
 
-import scala.collection.mutable
-import scala.collection.mutable.{HashMap, HashSet}
 import scala.concurrent.duration._
 import scala.io.Source
 import scala.jdk.CollectionConverters._
-import scala.reflect.ClassTag
 
 import org.json4s._
 import org.json4s.jackson.JsonMethods._
-import org.mockito.ArgumentMatchers.{any, eq => meq}
-import org.mockito.Mockito.{doNothing, mock, times, verify, when}
-import org.scalatest.{BeforeAndAfter, PrivateMethodTester}
-import org.scalatest.concurrent.Eventually
-import org.scalatest.matchers.must.Matchers
+import org.mockito.ArgumentMatchers.{eq => meq}
+import org.mockito.Mockito.{times, verify, when}
 import org.scalatest.matchers.should.Matchers._
 import org.scalatestplus.mockito.MockitoSugar.{mock => smock}
 import other.supplier.{CustomPersistenceEngine, CustomRecoveryModeFactory}
 
-import org.apache.spark.{SecurityManager, SparkConf, SparkFunSuite}
+import org.apache.spark.SparkConf
 import org.apache.spark.deploy._
 import org.apache.spark.deploy.DeployMessages._
 import org.apache.spark.internal.config._
 import org.apache.spark.internal.config.Deploy._
-import org.apache.spark.internal.config.Deploy.WorkerSelectionPolicy._
 import org.apache.spark.internal.config.UI._
 import org.apache.spark.internal.config.Worker._
 import org.apache.spark.io.LZ4CompressionCodec
@@ -53,121 +45,10 @@ import org.apache.spark.resource.{ResourceInformation, ResourceProfile, Resource
 import org.apache.spark.resource.ResourceProfile.DEFAULT_RESOURCE_PROFILE_ID
 import org.apache.spark.resource.ResourceUtils.{FPGA, GPU}
 import org.apache.spark.rpc.{RpcAddress, RpcEndpoint, RpcEndpointRef, RpcEnv}
-import org.apache.spark.serializer
 import org.apache.spark.serializer.KryoSerializer
 import org.apache.spark.util.Utils
 
-object MockWorker {
-  val counter = new AtomicInteger(10000)
-}
-
-class MockWorker(master: RpcEndpointRef, conf: SparkConf = new SparkConf) extends RpcEndpoint {
-  val seq = MockWorker.counter.incrementAndGet()
-  val id = seq.toString
-  override val rpcEnv: RpcEnv = RpcEnv.create("worker", "localhost", seq,
-    conf, new SecurityManager(conf))
-  val apps = new mutable.HashMap[String, String]()
-  val driverIdToAppId = new mutable.HashMap[String, String]()
-  def newDriver(driverId: String): RpcEndpointRef = {
-    val name = s"driver_${drivers.size}"
-    rpcEnv.setupEndpoint(name, new RpcEndpoint {
-      override val rpcEnv: RpcEnv = MockWorker.this.rpcEnv
-      override def receive: PartialFunction[Any, Unit] = {
-        case RegisteredApplication(appId, _) =>
-          apps(appId) = appId
-          driverIdToAppId(driverId) = appId
-      }
-    })
-  }
-
-  var decommissioned = false
-  var appDesc = DeployTestUtils.createAppDesc()
-  val drivers = mutable.HashSet[String]()
-  val driverResources = new mutable.HashMap[String, Map[String, Set[String]]]
-  val execResources = new mutable.HashMap[String, Map[String, Set[String]]]
-  val launchedExecutors = new mutable.HashMap[String, LaunchExecutor]
-  override def receive: PartialFunction[Any, Unit] = {
-    case RegisteredWorker(masterRef, _, _, _) =>
-      masterRef.send(WorkerLatestState(id, Nil, drivers.toSeq))
-    case l @ LaunchExecutor(_, appId, execId, _, _, _, _, resources_) =>
-      execResources(appId + "/" + execId) = resources_.map(r => (r._1, r._2.addresses.toSet))
-      launchedExecutors(appId + "/" + execId) = l
-    case LaunchDriver(driverId, desc, resources_) =>
-      drivers += driverId
-      driverResources(driverId) = resources_.map(r => (r._1, r._2.addresses.toSet))
-      master.send(RegisterApplication(appDesc, newDriver(driverId)))
-    case KillDriver(driverId) =>
-      master.send(DriverStateChanged(driverId, DriverState.KILLED, None))
-      drivers -= driverId
-      driverResources.remove(driverId)
-      driverIdToAppId.get(driverId) match {
-        case Some(appId) =>
-          apps.remove(appId)
-          master.send(UnregisterApplication(appId))
-        case None =>
-      }
-      driverIdToAppId.remove(driverId)
-    case DecommissionWorker =>
-      decommissioned = true
-  }
-}
-
-// This class is designed to handle the lifecycle of only one application.
-class MockExecutorLaunchFailWorker(master: Master, conf: SparkConf = new SparkConf)
-  extends MockWorker(master.self, conf) with Eventually {
-
-  val appRegistered = new CountDownLatch(1)
-  val launchExecutorReceived = new CountDownLatch(1)
-  val appIdsToLaunchExecutor = new mutable.HashSet[String]
-  var failedCnt = 0
-
-  override def receive: PartialFunction[Any, Unit] = {
-    case LaunchDriver(driverId, _, _) =>
-      master.self.send(RegisterApplication(appDesc, newDriver(driverId)))
-
-      // Below code doesn't make driver stuck, as newDriver opens another rpc endpoint for
-      // handling driver related messages. To simplify logic, we will block handling
-      // LaunchExecutor message until we validate registering app succeeds.
-      eventually(timeout(5.seconds)) {
-        // an app would be registered with Master once Driver set up
-        assert(apps.nonEmpty)
-        assert(master.idToApp.keySet.intersect(apps.keySet) == apps.keySet)
-      }
-
-      appRegistered.countDown()
-    case LaunchExecutor(_, appId, execId, _, _, _, _, _) =>
-      assert(appRegistered.await(10, TimeUnit.SECONDS))
-
-      if (failedCnt == 0) {
-        launchExecutorReceived.countDown()
-      }
-      assert(master.idToApp.contains(appId))
-      appIdsToLaunchExecutor += appId
-      failedCnt += 1
-      master.self.askSync(ExecutorStateChanged(appId, execId,
-        ExecutorState.FAILED, None, None))
-
-    case otherMsg => super.receive(otherMsg)
-  }
-}
-
-class MasterSuite extends SparkFunSuite
-  with Matchers with Eventually with PrivateMethodTester with BeforeAndAfter {
-
-  // regex to extract worker links from the master webui HTML
-  // groups represent URL and worker ID
-  val WORKER_LINK_RE = """<a href="(.+?)">\s*(worker-.+?)\s*</a>""".r
-
-  private var _master: Master = _
-
-  after {
-    if (_master != null) {
-      _master.rpcEnv.shutdown()
-      _master.rpcEnv.awaitTermination()
-      _master = null
-    }
-  }
-
+class MasterSuite extends MasterSuiteBase {
   test("can use a custom recovery mode factory") {
     val conf = new SparkConf(loadDefaults = false)
     conf.set(RECOVERY_MODE, "CUSTOM")
@@ -576,33 +457,6 @@ class MasterSuite extends SparkFunSuite
     }
   }
 
-  private def verifyWorkerUI(masterHtml: String, masterUrl: String,
-      reverseProxyUrl: String = ""): Unit = {
-    val workerLinks = (WORKER_LINK_RE findAllMatchIn masterHtml).toList
-    workerLinks.size should be (2)
-    workerLinks foreach {
-      case WORKER_LINK_RE(workerUrl, workerId) =>
-        workerUrl should be (s"$reverseProxyUrl/proxy/$workerId")
-        // there is no real front-end proxy as defined in $reverseProxyUrl
-        // construct url directly targeting the master
-        val url = s"$masterUrl/proxy/$workerId/"
-        System.setProperty("spark.ui.proxyBase", workerUrl)
-        val workerHtml = Utils
-          .tryWithResource(Source.fromURL(url))(_.getLines().mkString("\n"))
-        workerHtml should include ("Spark Worker at")
-        workerHtml should include ("Running Executors (0)")
-        verifyStaticResourcesServedByProxy(workerHtml, workerUrl)
-      case _ => fail()  // make sure we don't accidentially skip the tests
-    }
-  }
-
-  private def verifyStaticResourcesServedByProxy(html: String, proxyUrl: String): Unit = {
-    html should not include ("""href="/static""")
-    html should include (s"""href="$proxyUrl/static""")
-    html should not include ("""src="/static""")
-    html should include (s"""src="$proxyUrl/static""")
-  }
-
   test("basic scheduling - spread out") {
     basicScheduling(spreadOut = true)
   }
@@ -683,57 +537,6 @@ class MasterSuite extends SparkFunSuite
     scheduleExecutorsForAppWithMultiRPs(withMaxCores = true)
   }
 
-
-  private val workerSelectionPolicyTestCases = Seq(
-    (CORES_FREE_ASC, true, List("10001", "10002")),
-    (CORES_FREE_ASC, false, List("10001")),
-    (CORES_FREE_DESC, true, List("10002", "10003")),
-    (CORES_FREE_DESC, false, List("10003")),
-    (MEMORY_FREE_ASC, true, List("10001", "10003")),
-    (MEMORY_FREE_ASC, false, List("10001")),
-    (MEMORY_FREE_DESC, true, List("10002", "10003")),
-    (MEMORY_FREE_DESC, false, Seq("10002")),
-    (WORKER_ID, true, Seq("10001", "10002")),
-    (WORKER_ID, false, Seq("10001")))
-
-  workerSelectionPolicyTestCases.foreach { case (policy, spreadOut, expected) =>
-    test(s"SPARK-46881: scheduling with workerSelectionPolicy - $policy ($spreadOut)") {
-      val conf = new SparkConf()
-        .set(WORKER_SELECTION_POLICY.key, policy.toString)
-        .set(SPREAD_OUT_APPS.key, spreadOut.toString)
-        .set(UI_ENABLED.key, "false")
-        .set(Network.RPC_NETTY_DISPATCHER_NUM_THREADS, 1)
-        .set(Network.RPC_IO_THREADS, 1)
-      val master = makeAliveMaster(conf)
-
-      // Use different core and memory values to simplify the tests
-      MockWorker.counter.set(10000)
-      (1 to 3).foreach { idx =>
-        val worker = new MockWorker(master.self, conf)
-        worker.rpcEnv.setupEndpoint(s"worker-$idx", worker)
-        val workerReg = RegisterWorker(
-          worker.id,
-          "localhost",
-          worker.self.address.port,
-          worker.self,
-          4 + idx,
-          1280 * (if (idx < 2) idx else (6 - idx)),
-          "http://localhost:8080",
-          RpcAddress("localhost", 10000))
-        master.self.send(workerReg)
-        eventually(timeout(10.seconds)) {
-          assert(master.self.askSync[MasterStateResponse](RequestMasterState).workers.size === idx)
-        }
-      }
-
-      // An application with two executors
-      val appInfo = makeAppInfo(128, Some(2), Some(4))
-      master.registerApplication(appInfo)
-      startExecutorsOnWorkers(master)
-      assert(appInfo.executors.map(_._2.worker.id).toSeq.distinct.sorted === expected)
-    }
-  }
-
   test("SPARK-45174: scheduling with max drivers") {
     val master = makeMaster(new SparkConf().set(MAX_DRIVERS, 4))
     master.state = RecoveryState.ALIVE
@@ -770,334 +573,6 @@ class MasterSuite extends SparkFunSuite
 
   test("SPARK-46800: schedule not to spread out drivers") {
     verifyDrivers(false, 3, 0, 0)
-  }
-
-  private def verifyDrivers(spreadOut: Boolean, answer1: Int, answer2: Int, answer3: Int): Unit = {
-    val master = makeMaster(new SparkConf().set(SPREAD_OUT_DRIVERS, spreadOut))
-    val worker1 = makeWorkerInfo(4096, 10)
-    val worker2 = makeWorkerInfo(4096, 10)
-    val worker3 = makeWorkerInfo(4096, 10)
-    master.state = RecoveryState.ALIVE
-    master.workers += worker1
-    master.workers += worker2
-    master.workers += worker3
-    val drivers = getDrivers(master)
-    val waitingDrivers = master.invokePrivate(_waitingDrivers())
-
-    master.invokePrivate(_schedule())
-    assert(drivers.size === 0 && waitingDrivers.size === 0)
-
-    val command = Command("", Seq.empty, Map.empty, Seq.empty, Seq.empty, Seq.empty)
-    val desc = DriverDescription("", 1, 1, false, command)
-    (1 to 3).foreach { i =>
-      val driver = new DriverInfo(0, "driver" + i, desc, new Date())
-      waitingDrivers += driver
-      drivers.add(driver)
-    }
-    assert(drivers.size === 3 && waitingDrivers.size === 3)
-    master.invokePrivate(_schedule())
-    assert(drivers.size === 3 && waitingDrivers.size === 0)
-    assert(worker1.drivers.size === answer1)
-    assert(worker2.drivers.size === answer2)
-    assert(worker3.drivers.size === answer3)
-  }
-
-  private def scheduleExecutorsForAppWithMultiRPs(withMaxCores: Boolean): Unit = {
-    val appInfo: ApplicationInfo = if (withMaxCores) {
-      makeAppInfo(
-        1024, maxCores = Some(30), initialExecutorLimit = Some(0))
-    } else {
-      makeAppInfo(
-        1024, maxCores = None, initialExecutorLimit = Some(0))
-    }
-
-    val master = makeAliveMaster()
-    val conf = new SparkConf()
-    val workers = (1 to 4).map { idx =>
-      val worker = new MockWorker(master.self, conf)
-      worker.rpcEnv.setupEndpoint(s"worker-$idx", worker)
-      val workerReg = RegisterWorker(
-        worker.id,
-        "localhost",
-        worker.self.address.port,
-        worker.self,
-        10,
-        4096,
-        "http://localhost:8080",
-        RpcAddress("localhost", 10000))
-      master.self.send(workerReg)
-      worker
-    }
-
-    // Register app and schedule.
-    master.registerApplication(appInfo)
-    startExecutorsOnWorkers(master)
-    assert(appInfo.executors.isEmpty)
-
-    // Request executors with multiple resource profile.
-    // rp1 with 15 cores per executor, rp2 with 8192MB memory per executor, no worker can
-    // fulfill the resource requirement.
-    val rp1 = DeployTestUtils.createResourceProfile(Some(2048), Map.empty, Some(15))
-    val rp2 = DeployTestUtils.createResourceProfile(Some(8192), Map.empty, Some(5))
-    val rp3 = DeployTestUtils.createResourceProfile(Some(2048), Map.empty, Some(5))
-    val rp4 = DeployTestUtils.createResourceProfile(Some(2048), Map.empty, Some(10))
-    val requests = Map(
-      appInfo.desc.defaultProfile -> 1,
-      rp1 -> 1,
-      rp2 -> 1,
-      rp3 -> 1,
-      rp4 -> 2
-    )
-    eventually(timeout(10.seconds)) {
-      master.self.askSync[Boolean](RequestExecutors(appInfo.id, requests))
-      assert(appInfo.executors.size === workers.map(_.launchedExecutors.size).sum)
-    }
-
-    if (withMaxCores) {
-      assert(appInfo.executors.size === 3)
-      assert(appInfo.getOrUpdateExecutorsForRPId(DEFAULT_RESOURCE_PROFILE_ID).size === 1)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp1.id).size === 0)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp2.id).size === 0)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp3.id).size === 1)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp4.id).size === 1)
-    } else {
-      assert(appInfo.executors.size === 4)
-      assert(appInfo.getOrUpdateExecutorsForRPId(DEFAULT_RESOURCE_PROFILE_ID).size === 1)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp1.id).size === 0)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp2.id).size === 0)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp3.id).size === 1)
-      assert(appInfo.getOrUpdateExecutorsForRPId(rp4.id).size === 2)
-    }
-
-    // Verify executor information.
-    val executorForRp3 = appInfo.executors(appInfo.getOrUpdateExecutorsForRPId(rp3.id).head)
-    assert(executorForRp3.cores === 5)
-    assert(executorForRp3.memory === 2048)
-    assert(executorForRp3.rpId === rp3.id)
-
-    // Verify LaunchExecutor message.
-    val launchExecutorMsg = workers
-      .find(_.id === executorForRp3.worker.id)
-      .map(_.launchedExecutors(appInfo.id + "/" + executorForRp3.id))
-      .get
-    assert(launchExecutorMsg.cores === 5)
-    assert(launchExecutorMsg.memory === 2048)
-    assert(launchExecutorMsg.rpId === rp3.id)
-  }
-
-  private def basicScheduling(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo = makeAppInfo(1024)
-    val scheduledCores = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    assert(scheduledCores === Array(10, 10, 10))
-  }
-
-  private def basicSchedulingWithMoreMemory(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo = makeAppInfo(3072)
-    val scheduledCores = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    assert(scheduledCores === Array(10, 10, 10))
-  }
-
-  private def schedulingWithMaxCores(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo1 = makeAppInfo(1024, maxCores = Some(8))
-    val appInfo2 = makeAppInfo(1024, maxCores = Some(16))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo1, workerInfos, spreadOut)
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo2, workerInfos, spreadOut)
-    if (spreadOut) {
-      assert(scheduledCores1 === Array(3, 3, 2))
-      assert(scheduledCores2 === Array(6, 5, 5))
-    } else {
-      assert(scheduledCores1 === Array(8, 0, 0))
-      assert(scheduledCores2 === Array(10, 6, 0))
-    }
-  }
-
-  private def schedulingWithCoresPerExecutor(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo1 = makeAppInfo(1024, coresPerExecutor = Some(2))
-    val appInfo2 = makeAppInfo(256, coresPerExecutor = Some(2))
-    val appInfo3 = makeAppInfo(256, coresPerExecutor = Some(3))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo1, workerInfos, spreadOut)
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo2, workerInfos, spreadOut)
-    val scheduledCores3 = scheduleExecutorsOnWorkers(master, appInfo3, workerInfos, spreadOut)
-    assert(scheduledCores1 === Array(8, 8, 8)) // 4 * 2 because of memory limits
-    assert(scheduledCores2 === Array(10, 10, 10)) // 5 * 2
-    assert(scheduledCores3 === Array(9, 9, 9)) // 3 * 3
-  }
-
-  // Sorry for the long method name!
-  private def schedulingWithCoresPerExecutorAndMaxCores(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo1 = makeAppInfo(256, coresPerExecutor = Some(2), maxCores = Some(4))
-    val appInfo2 = makeAppInfo(256, coresPerExecutor = Some(2), maxCores = Some(20))
-    val appInfo3 = makeAppInfo(256, coresPerExecutor = Some(3), maxCores = Some(20))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo1, workerInfos, spreadOut)
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo2, workerInfos, spreadOut)
-    val scheduledCores3 = scheduleExecutorsOnWorkers(master, appInfo3, workerInfos, spreadOut)
-    if (spreadOut) {
-      assert(scheduledCores1 === Array(2, 2, 0))
-      assert(scheduledCores2 === Array(8, 6, 6))
-      assert(scheduledCores3 === Array(6, 6, 6))
-    } else {
-      assert(scheduledCores1 === Array(4, 0, 0))
-      assert(scheduledCores2 === Array(10, 10, 0))
-      assert(scheduledCores3 === Array(9, 9, 0))
-    }
-  }
-
-  private def schedulingWithExecutorLimit(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo = makeAppInfo(256)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 0))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 2))
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 5))
-    val scheduledCores3 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    assert(scheduledCores1 === Array(0, 0, 0))
-    assert(scheduledCores2 === Array(10, 10, 0))
-    assert(scheduledCores3 === Array(10, 10, 10))
-  }
-
-  private def schedulingWithExecutorLimitAndMaxCores(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo = makeAppInfo(256, maxCores = Some(16))
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 0))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 2))
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 5))
-    val scheduledCores3 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    assert(scheduledCores1 === Array(0, 0, 0))
-    if (spreadOut) {
-      assert(scheduledCores2 === Array(8, 8, 0))
-      assert(scheduledCores3 === Array(6, 5, 5))
-    } else {
-      assert(scheduledCores2 === Array(10, 6, 0))
-      assert(scheduledCores3 === Array(10, 6, 0))
-    }
-  }
-
-  private def schedulingWithExecutorLimitAndCoresPerExecutor(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo = makeAppInfo(256, coresPerExecutor = Some(4))
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 0))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 2))
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 5))
-    val scheduledCores3 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    assert(scheduledCores1 === Array(0, 0, 0))
-    if (spreadOut) {
-      assert(scheduledCores2 === Array(4, 4, 0))
-    } else {
-      assert(scheduledCores2 === Array(8, 0, 0))
-    }
-    assert(scheduledCores3 === Array(8, 8, 4))
-  }
-
-  // Everything being: executor limit + cores per executor + max cores
-  private def schedulingWithEverything(spreadOut: Boolean): Unit = {
-    val master = makeMaster()
-    val appInfo = makeAppInfo(256, coresPerExecutor = Some(4), maxCores = Some(18))
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 0))
-    val scheduledCores1 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 2))
-    val scheduledCores2 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    appInfo.requestExecutors(Map(appInfo.desc.defaultProfile -> 5))
-    val scheduledCores3 = scheduleExecutorsOnWorkers(master, appInfo, workerInfos, spreadOut)
-    assert(scheduledCores1 === Array(0, 0, 0))
-    if (spreadOut) {
-      assert(scheduledCores2 === Array(4, 4, 0))
-      assert(scheduledCores3 === Array(8, 4, 4))
-    } else {
-      assert(scheduledCores2 === Array(8, 0, 0))
-      assert(scheduledCores3 === Array(8, 8, 0))
-    }
-  }
-
-  // ==========================================
-  // | Utility methods and fields for testing |
-  // ==========================================
-
-  private val _schedule = PrivateMethod[Unit](Symbol("schedule"))
-  private val _scheduleExecutorsOnWorkers =
-    PrivateMethod[Array[Int]](Symbol("scheduleExecutorsOnWorkers"))
-  private val _startExecutorsOnWorkers =
-    PrivateMethod[Unit](Symbol("startExecutorsOnWorkers"))
-  private val _drivers = PrivateMethod[HashSet[DriverInfo]](Symbol("drivers"))
-  private val _waitingDrivers =
-    PrivateMethod[mutable.ArrayBuffer[DriverInfo]](Symbol("waitingDrivers"))
-  private val _state = PrivateMethod[RecoveryState.Value](Symbol("state"))
-  private val _newDriverId = PrivateMethod[String](Symbol("newDriverId"))
-  private val _newApplicationId = PrivateMethod[String](Symbol("newApplicationId"))
-  private val _createApplication = PrivateMethod[ApplicationInfo](Symbol("createApplication"))
-  private val _persistenceEngine = PrivateMethod[PersistenceEngine](Symbol("persistenceEngine"))
-
-  private val workerInfo = makeWorkerInfo(4096, 10)
-  private val workerInfos = Array(workerInfo, workerInfo, workerInfo)
-
-  private def makeMaster(conf: SparkConf = new SparkConf): Master = {
-    assert(_master === null, "Some Master's RpcEnv is leaked in tests")
-    val securityMgr = new SecurityManager(conf)
-    val rpcEnv = RpcEnv.create(Master.SYSTEM_NAME, "localhost", 0, conf, securityMgr)
-    _master = new Master(rpcEnv, rpcEnv.address, 0, securityMgr, conf)
-    _master
-  }
-
-  def makeAliveMaster(conf: SparkConf = new SparkConf): Master = {
-    val master = makeMaster(conf)
-    master.rpcEnv.setupEndpoint(Master.ENDPOINT_NAME, master)
-    eventually(timeout(10.seconds)) {
-      val masterState = master.self.askSync[MasterStateResponse](RequestMasterState)
-      assert(masterState.status === RecoveryState.ALIVE, "Master is not alive")
-    }
-    master
-  }
-
-  private def makeAppInfo(
-      memoryPerExecutorMb: Int,
-      coresPerExecutor: Option[Int] = None,
-      maxCores: Option[Int] = None,
-      customResources: Map[String, Int] = Map.empty,
-      initialExecutorLimit: Option[Int] = None): ApplicationInfo = {
-    val rp = DeployTestUtils.createDefaultResourceProfile(
-      memoryPerExecutorMb, customResources, coresPerExecutor)
-
-    val desc = new ApplicationDescription(
-      "test", maxCores, null, "", rp, None, None, initialExecutorLimit)
-    val appId = System.nanoTime().toString
-    val endpointRef = mock(classOf[RpcEndpointRef])
-    val mockAddress = mock(classOf[RpcAddress])
-    when(endpointRef.address).thenReturn(mockAddress)
-    doNothing().when(endpointRef).send(any())
-    new ApplicationInfo(0, appId, desc, new Date, endpointRef, Int.MaxValue)
-  }
-
-  private def makeWorkerInfo(memoryMb: Int, cores: Int): WorkerInfo = {
-    val workerId = System.nanoTime().toString
-    val endpointRef = mock(classOf[RpcEndpointRef])
-    val mockAddress = mock(classOf[RpcAddress])
-    when(endpointRef.address).thenReturn(mockAddress)
-    new WorkerInfo(workerId, "host", 100, cores, memoryMb,
-      endpointRef, "http://localhost:80", Map.empty)
-  }
-
-  // Schedule executors for default resource profile.
-  private def scheduleExecutorsOnWorkers(
-      master: Master,
-      appInfo: ApplicationInfo,
-      workerInfos: Array[WorkerInfo],
-      spreadOut: Boolean): Array[Int] = {
-    val defaultResourceDesc = appInfo.getResourceDescriptionForRpId(DEFAULT_RESOURCE_PROFILE_ID)
-    master.invokePrivate(_scheduleExecutorsOnWorkers(
-      appInfo, DEFAULT_RESOURCE_PROFILE_ID, defaultResourceDesc, workerInfos, spreadOut))
-  }
-
-  private def startExecutorsOnWorkers(master: Master): Unit = {
-    master.invokePrivate(_startExecutorsOnWorkers())
   }
 
   test("SPARK-13604: Master should ask Worker kill unknown executors and drivers") {
@@ -1208,56 +683,6 @@ class MasterSuite extends SparkFunSuite
         master.rpcEnv.shutdown()
       }
     }
-  }
-
-  def testWorkerDecommissioning(
-      numWorkers: Int,
-      numWorkersExpectedToDecom: Int,
-      hostnames: Seq[String]): Unit = {
-    val conf = new SparkConf()
-    val master = makeAliveMaster(conf)
-    val workers = (1 to numWorkers).map { idx =>
-      val worker = new MockWorker(master.self, conf)
-      worker.rpcEnv.setupEndpoint(s"worker-$idx", worker)
-      val workerReg = RegisterWorker(
-        worker.id,
-        "localhost",
-        worker.self.address.port,
-        worker.self,
-        10,
-        1024,
-        "http://localhost:8080",
-        RpcAddress("localhost", 10000))
-      master.self.send(workerReg)
-      worker
-    }
-
-    eventually(timeout(10.seconds)) {
-      val masterState = master.self.askSync[MasterStateResponse](RequestMasterState)
-      assert(masterState.workers.length === numWorkers)
-      assert(masterState.workers.forall(_.state == WorkerState.ALIVE))
-      assert(masterState.workers.map(_.id).toSet == workers.map(_.id).toSet)
-    }
-
-    val decomWorkersCount = master.self.askSync[Integer](DecommissionWorkersOnHosts(hostnames))
-    assert(decomWorkersCount === numWorkersExpectedToDecom)
-
-    // Decommissioning is actually async ... wait for the workers to actually be decommissioned by
-    // polling the master's state.
-    eventually(timeout(30.seconds)) {
-      val masterState = master.self.askSync[MasterStateResponse](RequestMasterState)
-      assert(masterState.workers.length === numWorkers)
-      val workersActuallyDecomed = masterState.workers
-        .filter(_.state == WorkerState.DECOMMISSIONED).map(_.id)
-      val decommissionedWorkers = workers.filter(w => workersActuallyDecomed.contains(w.id))
-      assert(workersActuallyDecomed.length === numWorkersExpectedToDecom)
-      assert(decommissionedWorkers.forall(_.decommissioned))
-    }
-
-    // Decommissioning a worker again should return the same answer since we want this call to be
-    // idempotent.
-    val decomWorkersCountAgain = master.self.askSync[Integer](DecommissionWorkersOnHosts(hostnames))
-    assert(decomWorkersCountAgain === numWorkersExpectedToDecom)
   }
 
   test("All workers on a host should be decommissioned") {
@@ -1462,14 +887,6 @@ class MasterSuite extends SparkFunSuite
       Seq(ResourceRequirement(FPGA, 3), ResourceRequirement(GPU, 3)))
   }
 
-  private def getDrivers(master: Master): HashSet[DriverInfo] = {
-    master.invokePrivate(_drivers())
-  }
-
-  private def getState(master: Master): RecoveryState.Value = {
-    master.invokePrivate(_state())
-  }
-
   test("SPARK-45753: Support driver id pattern") {
     val master = makeMaster(new SparkConf().set(DRIVER_ID_PATTERN, "my-driver-%2$05d"))
     val submitDate = new Date()
@@ -1575,32 +992,4 @@ class MasterSuite extends SparkFunSuite
     assert(master.workers.nonEmpty)
     assert(master.idToWorker.nonEmpty)
   }
-}
-
-private class FakeRecoveryModeFactory(conf: SparkConf, ser: serializer.Serializer)
-    extends StandaloneRecoveryModeFactory(conf, ser) {
-  import FakeRecoveryModeFactory.persistentData
-
-  override def createPersistenceEngine(): PersistenceEngine = new PersistenceEngine {
-
-    override def unpersist(name: String): Unit = {
-      persistentData.remove(name)
-    }
-
-    override def persist(name: String, obj: Object): Unit = {
-      persistentData(name) = obj
-    }
-
-    override def read[T: ClassTag](prefix: String): Seq[T] = {
-      persistentData.filter(_._1.startsWith(prefix)).map(_._2.asInstanceOf[T]).toSeq
-    }
-  }
-
-  override def createLeaderElectionAgent(master: LeaderElectable): LeaderElectionAgent = {
-    new MonarchyLeaderAgent(master)
-  }
-}
-
-private object FakeRecoveryModeFactory {
-  val persistentData = new HashMap[String, Object]()
 }
